+++ src/pass/ip_addr.rs
@@ -5,32 +5,43 @@ use std::net::IpAddr;
 
 /// Run all passes with this `IpAddr`.
 pub fn run(ip: &IpAddr, config: &Config) {
-    pass_sequence!(ip, config, "ip"; general, specific, conversions);
+    pass_sequence!(ip, config, "ip"; information, conversions);
 }
 
-/// Prints general information about this `IpAddr`
-pub fn general(ip: &IpAddr) {
-    match ip {
-        IpAddr::V4(_) => info!("type", "IPv4"),
-        IpAddr::V6(_) => info!("type", "IPv6"),
-    }
+/// Prints information about this `IpAddr`
+pub fn information(ip: &IpAddr) {
+    let other = "other".cyan().bold().to_string();
 
-    info!("unspecified?", ip.is_unspecified());
-    info!("loopback?", ip.is_loopback());
-    info!("multicast?", ip.is_multicast());
-}
-
-/// Prints information about this `IpAddr` specific to its version
-pub fn specific(ip: &IpAddr) {
     match ip {
         IpAddr::V4(v) => {
+            info!("type", "IPv4");
+
+            info!("kind", match v {
+                _ if v.is_unspecified() => "unspecified",
+                _ if v.is_loopback() => "loopback",
+                _ if v.is_multicast() => "multicast",
+
+                _ if v.is_private() => "private",
+                _ if v.is_link_local() => "link-local",
+                _ if v.is_broadcast() => "broadcast",
+                _ if v.is_documentation() => "documentation",
+
+                _ => &other,
+            });
+
             info!("octets", format!("{:?}", v.octets()));
-            info!("private?", v.is_private());
-            info!("link-local?", v.is_link_local());
-            info!("broadcast?", v.is_broadcast());
-            info!("documentation?", v.is_documentation());
         },
         IpAddr::V6(v) => {
+            info!("type", "IPv6");
+
+            info!("kind", match v {
+                _ if v.is_unspecified() => "unspecified",
+                _ if v.is_loopback() => "loopback",
+                _ if v.is_multicast() => "multicast",
+
+                _ => &other,
+            });
+
             info!("octets", format!("{:?}", v.octets()));
             info!("segments", format!("{:?}", v.segments()));
         },
@@ -41,8 +52,8 @@ pub fn specific(ip: &IpAddr) {
 pub fn conversions(ip: &IpAddr) {
     match ip {
         IpAddr::V4(v) => {
-            info!("ipv6-compatible", v.to_ipv6_compatible());
-            info!("ipv6-mapped", v.to_ipv6_mapped());
+            info!("ipv6 compatible", v.to_ipv6_compatible());
+            info!("ipv6 mapped", v.to_ipv6_mapped());
         },
         IpAddr::V6(v) => {
             match v.to_ipv4() {