+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.1] - 2018-12-31
+### Changed
+- Merged `ip_addr -> {general, specific}` into `ip_addr -> information`
+- Merged `ip_addr -> {general, specific} -> {unspecified?, loopback?, multicast?, private?, link-local?, broadcast?, documentation?}` into `ip_addr -> information -> kind` since at most one of these properties should be `true` for a given IP address
+- Renamed `ip_addr -> conversions -> {ipv6-compatible, ipv6-mapped}` to `ip_addr -> conversions -> {ipv6 compatible, ipv6 mapped}`
+
 ## [0.7.0] - 2018-12-30
 ### Added
 - Parsing of and information for IP addresses (via `IpAddr`), see `README.md` for more information!